GraphingCalculator - Update commit - Hour 2
Added more buttons for the graphing feature, and added a domain restriction as well (maths domain lol). Also made it plot 5000 values of x within the domain using a for loop of this equation: MinumumDomain + [(MaximumDomain - MinumumDomain) * n] / 5000. It gets 5000 values of x, subs it into the equation to get 5000 values of y, and 5000 pairs of xy coodinates which are plotted using chart.js